Improve maintainability
